64-bit fixes

+++ src/tree/itemise.cpp
@@ -2,5 +2,5 @@
 
 const char *Itemiser::dstart;
 const char *Itemiser::dend;
-vector<int> Itemiser::v;
+vector<uintptr_t> Itemiser::v;
